Filip/milestone 3/set up

+++ TravelAgency-persistence/src/main/java/cz/muni/fi/travelAgency/dao/TripDaoImpl.java
@@ -1,8 +1,10 @@
 package cz.muni.fi.travelAgency.dao;
 
+import cz.muni.fi.travelAgency.entities.Excursion;
+import cz.muni.fi.travelAgency.entities.Reservation;
 import cz.muni.fi.travelAgency.entities.Trip;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -15,11 +17,16 @@ import java.util.Collection;
  * @author Simona Raucinova
  */
 @Repository
-@Transactional
 public class TripDaoImpl implements TripDao {
 
     @PersistenceContext
     private EntityManager em;
+
+    @Autowired
+    private ExcursionDao excursionDao;
+
+    @Autowired
+    private ReservationDao reservationDao;
 
     @Override
     public void create(Trip trip) {
@@ -63,6 +70,20 @@ public class TripDaoImpl implements TripDao {
 
     @Override
     public void remove(Trip trip) {
+        if (trip == null) {
+            throw new IllegalArgumentException("Tried to delete NULL trip");
+        }
+        if (findById(trip.getId()) == null) {
+            throw new IllegalArgumentException("Tried to delete trip that was not saved");
+        }
+        Collection<Reservation> reservations = trip.getReservations();
+        Collection<Excursion> excursions = trip.getExcursions();
+        for (Excursion excursion : excursions) {
+            excursionDao.remove(excursion);
+        }
+        for (Reservation reservation : reservations) {
+            reservationDao.remove(reservation);
+        }
         em.remove(em.merge(trip));
     }
 }
